backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3>=1.40.0
openai>=1.50.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.3.4
websockets==12.0
httpx>=0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
